Feature/authorization

+++ app/Http/Controllers/CategoryController.php
@@ -2,21 +2,29 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Account;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
-
 use App\Models\Category;
 use App\Models\Idea;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\Validator;
 
 class CategoryController extends Controller
 {
     public function index(){
         $categories = Category::paginate(10)->sortByDesc('id');
-        return view('category', compact('categories'))->with('i', (request()-> input('page', 1) -1)*10);
+        if (Auth::guard('account')->user()->role == Account::ACCOUNT_ADMIN) {
+            return view('admin.category', compact('categories'))->with('i', (request()-> input('page', 1) -1)*10);
+        }
+        return view('user.category', compact('categories'))->with('i', (request()-> input('page', 1) -1)*10);
     }
 
     public function store(Request $request){
-
+        $request->validate([
+            'category_name' => 'required|max:255'
+        ]);
         $cate = Category::create([
             'category_name' => $request->category_name,
             'first_closure_date' => Carbon::now(),
@@ -30,10 +38,15 @@ class CategoryController extends Controller
     }
 
     public function update(Request $request, $id){
-
+//        if (! Gate::allows('edit-cate', $account)) {
+//            return redirect()->back()->with('error','You can not edit category');
+//        }
         $cate = Category::find($id);
 
         if($cate){
+            $request->validate([
+                'category_name' => 'required'
+            ]);
             $update_cate = $cate->update([
                 'category_name' => $request->input('category_name'),
             ]);
@@ -44,8 +57,14 @@ class CategoryController extends Controller
         }
     }
 
-    public function destroy(Request $request, $id){
+    public function destroy(Request $request, $id, Category $category){
+//        if (! Gate::allows('delete-cate', $category)) {
+//            return redirect()->back()->with('error','You can not delete category');
+//        }
         $cate = Category::find($id);
+        if (count($cate->idea) != 0){
+            return redirect()->back()->with('error','Delete Category Not Success');
+        }
         if($cate){
             $delete_cate = $cate->delete();
